pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nfs-watcher-uploader"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "azure-storage-blob>=12.23.0",
    "azure-identity>=1.17.0",
    "uuid-utils>=0.9.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-http>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "anyio>=4.0.0",
    "pydantic-settings>=2.0.0",
]

[project.scripts]
nfs-watcher-uploader = "app.main:run"

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pyright>=1.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
